thorium_node: use a 64-bit integer for the tick number

+++ engine/thorium/node.h
@@ -206,7 +206,7 @@ struct thorium_node {
     int provided;
 
     int name;
-    int tick;
+    uint64_t tick;
     int nodes;
     int threads;
